Move functionality around, hopefully where it's best placed.
