+++ libs/pavex/src/language/resolved_type.rs
@@ -1,6 +1,7 @@
 use std::fmt::Write;
 use std::fmt::{Debug, Display, Formatter};
 
+use ahash::{HashMap, HashMapExt};
 use anyhow::Context;
 use bimap::BiHashMap;
 use guppy::PackageId;
@@ -19,6 +20,199 @@ pub enum ResolvedType {
 
 impl ResolvedType {
     pub const UNIT_TYPE: ResolvedType = ResolvedType::Tuple(Tuple { elements: vec![] });
+
+    /// Replace unassigned generic type parameters in `templated_type` with the concrete generic type
+    /// parameters defined in `bindings`.
+    ///
+    /// This function can also be used to _partially_ bind the unassigned generic type parameters in
+    /// `t`. You are not required to bind all of them.
+    pub fn bind_generic_type_parameters(
+        &self,
+        bindings: &HashMap<NamedTypeGeneric, ResolvedType>,
+    ) -> ResolvedType {
+        match self {
+            ResolvedType::ResolvedPath(t) => {
+                let mut bound_generics = Vec::with_capacity(t.generic_arguments.len());
+                for generic in &t.generic_arguments {
+                    let bound_generic = match generic {
+                        GenericArgument::UnassignedTypeParameter(name) => {
+                            if let Some(bound_type) = bindings.get(name) {
+                                GenericArgument::AssignedTypeParameter(bound_type.clone())
+                            } else {
+                                generic.to_owned()
+                            }
+                        }
+                        GenericArgument::AssignedTypeParameter(t) => {
+                            GenericArgument::AssignedTypeParameter(
+                                t.bind_generic_type_parameters(bindings),
+                            )
+                        }
+                        GenericArgument::Lifetime(_) => generic.to_owned(),
+                    };
+                    bound_generics.push(bound_generic);
+                }
+                ResolvedType::ResolvedPath(ResolvedPathType {
+                    package_id: t.package_id.clone(),
+                    // Should we set this to `None`?
+                    rustdoc_id: t.rustdoc_id.clone(),
+                    base_type: t.base_type.clone(),
+                    generic_arguments: bound_generics,
+                })
+            }
+            ResolvedType::Reference(r) => ResolvedType::Reference(TypeReference {
+                is_mutable: r.is_mutable,
+                inner: Box::new(r.inner.bind_generic_type_parameters(bindings)),
+                is_static: r.is_static,
+            }),
+            ResolvedType::Tuple(t) => {
+                let mut bound_elements = Vec::with_capacity(t.elements.len());
+                for inner in &t.elements {
+                    bound_elements.push(inner.bind_generic_type_parameters(bindings));
+                }
+                ResolvedType::Tuple(Tuple {
+                    elements: bound_elements,
+                })
+            }
+            ResolvedType::ScalarPrimitive(s) => ResolvedType::ScalarPrimitive(s.clone()),
+            ResolvedType::Slice(s) => ResolvedType::Slice(Slice {
+                element_type: Box::new(s.element_type.bind_generic_type_parameters(bindings)),
+            }),
+        }
+    }
+
+    /// Check if a type can be considered a "specialization" of another, with respect to their generic parameters.
+    ///
+    /// I.e. if by replacing the unassigned generic type parameters of `templated_type` with the
+    /// concrete generic type parameters of `concrete_type`, `templated_type` would be equal to `concrete_type`.
+    ///
+    /// If possible, this function will return a map associating each unassigned generic parameter
+    /// in `templated_type` with the type it must be set to in order to match `concrete_type`.
+    /// If impossible, this function will return `None`.
+    #[tracing::instrument(level = "trace", ret)]
+    pub fn is_a_template_for(
+        &self,
+        concrete_type: &ResolvedType,
+    ) -> Option<HashMap<NamedTypeGeneric, ResolvedType>> {
+        let mut bindings = HashMap::new();
+        if self._is_a_template_for(concrete_type, &mut bindings) {
+            Some(bindings)
+        } else {
+            None
+        }
+    }
+
+    #[tracing::instrument(level = "trace", ret)]
+    fn _is_a_template_for(
+        &self,
+        concrete_type: &ResolvedType,
+        bindings: &mut HashMap<NamedTypeGeneric, ResolvedType>,
+    ) -> bool {
+        if concrete_type == self {
+            return true;
+        }
+        use ResolvedType::*;
+        match (concrete_type, self) {
+            (ResolvedPath(concrete_path), ResolvedPath(templated_path)) => {
+                templated_path._is_a_resolved_path_type_template_for(concrete_path, bindings)
+            }
+            (Slice(concrete_slice), Slice(templated_slice)) => templated_slice
+                .element_type
+                ._is_a_template_for(&concrete_slice.element_type, bindings),
+            (Reference(concrete_reference), Reference(templated_reference)) => templated_reference
+                .inner
+                ._is_a_template_for(&concrete_reference.inner, bindings),
+            (Tuple(concrete_tuple), Tuple(templated_tuple)) => {
+                if concrete_tuple.elements.len() != templated_tuple.elements.len() {
+                    return false;
+                }
+                concrete_tuple
+                    .elements
+                    .iter()
+                    .zip(templated_tuple.elements.iter())
+                    .all(|(concrete_type, templated_type)| {
+                        templated_type._is_a_template_for(concrete_type, bindings)
+                    })
+            }
+            (ScalarPrimitive(concrete_primitive), ScalarPrimitive(templated_primitive)) => {
+                concrete_primitive == templated_primitive
+            }
+            (_, _) => false,
+        }
+    }
+}
+
+impl ResolvedPathType {
+    fn _is_a_resolved_path_type_template_for(
+        &self,
+        concrete_type: &ResolvedPathType,
+        bindings: &mut HashMap<NamedTypeGeneric, ResolvedType>,
+    ) -> bool {
+        // We destructure ALL fields to make sure that the compiler reminds us to update
+        // this function if we add new fields to `ResolvedPathType`.
+        let ResolvedPathType {
+            package_id: concrete_package_id,
+            rustdoc_id: _,
+            base_type: concrete_base_type,
+            generic_arguments: concrete_generic_arguments,
+        } = concrete_type;
+        let ResolvedPathType {
+            package_id: templated_package_id,
+            rustdoc_id: _,
+            base_type: templated_base_type,
+            generic_arguments: templated_generic_arguments,
+        } = self;
+        if concrete_package_id != templated_package_id
+            || concrete_base_type != templated_base_type
+            || concrete_generic_arguments.len() != templated_generic_arguments.len()
+        {
+            return false;
+        }
+        for (concrete_arg, templated_arg) in concrete_generic_arguments
+            .iter()
+            .zip(templated_generic_arguments.iter())
+        {
+            use GenericArgument::*;
+            match (concrete_arg, templated_arg) {
+                (
+                    AssignedTypeParameter(concrete_arg_type),
+                    AssignedTypeParameter(templated_arg_type),
+                ) => {
+                    if !templated_arg_type._is_a_template_for(concrete_arg_type, bindings) {
+                        return false;
+                    }
+                }
+                (AssignedTypeParameter(assigned), UnassignedTypeParameter(unassigned)) => {
+                    // The unassigned type parameter can be assigned to the concrete type
+                    // we expect, so it is a specialization.
+                    let previous_assignment = bindings.insert(unassigned.clone(), assigned.clone());
+                    if let Some(previous_assignment) = previous_assignment {
+                        if &previous_assignment != assigned {
+                            tracing::trace!(
+                            "Type parameter `{:?}` was already assigned to `{:?}` but is now being assigned to `{:?}`",
+                            unassigned,
+                            previous_assignment,
+                            assigned
+                        );
+                            return false;
+                        }
+                    }
+                }
+                (Lifetime(_), Lifetime(_)) => {
+                    // Lifetimes are not relevant for specialization (yet).
+                }
+                (UnassignedTypeParameter(unassigned), _) => {
+                    // You are not allowed to specialize a type with an unassigned type parameter.
+                    unreachable!("Unassigned type parameter (`{:?}`) in the 'concrete' type (`{:?}`) when checking for specialization", unassigned, concrete_type);
+                }
+                (AssignedTypeParameter(_), Lifetime(_))
+                | (Lifetime(_), UnassignedTypeParameter(_))
+                | (Lifetime(_), AssignedTypeParameter(_)) => {
+                    return false;
+                }
+            }
+        }
+        true
+    }
 }
 
 #[derive(serde::Serialize, serde::Deserialize, Eq, PartialEq, Hash, Clone)]
