+++ libs/pavex/src/web/analyses/call_graph/request_handler.rs
@@ -1,6 +1,8 @@
 use pavex_builder::Lifecycle;
 
-use crate::web::analyses::call_graph::{build_call_graph, CallGraph, NumberOfAllowedInvocations};
+use crate::web::analyses::call_graph::{
+    build_call_graph, debug_dot, CallGraph, NumberOfAllowedInvocations,
+};
 use crate::web::analyses::components::{ComponentDb, ComponentId};
 use crate::web::analyses::computations::ComputationDb;
 use crate::web::analyses::constructibles::ConstructibleDb;
@@ -30,6 +32,8 @@ pub(crate) fn handler_call_graph(
         constructible_db,
         lifecycle2invocations,
     );
+
+    println!("{}", debug_dot(&call_graph, &component_db, &computation_db));
 
     CallGraph {
         call_graph,
