+++ libs/pavex_cli/tests/ui_tests/output_type_of_constructors_can_contain_generic_parameters/lib.rs
@@ -1,10 +1,10 @@
 use pavex_builder::{f, router::GET, Blueprint, Lifecycle};
 
-pub fn json<T>() -> Json<T> {
-    todo!()
-}
-
-pub struct Json<T>(T);
+// pub fn json<T>() -> Json<T> {
+//     todo!()
+// }
+//
+// pub struct Json<T>(T);
 
 pub fn fallible<T>() -> Result<Form<T>, FallibleError> {
     todo!()
@@ -20,9 +20,9 @@ pub struct Form<T>(T);
 
 // The generic parameters of all inputs types are fully specified!
 pub fn handler(
-    json: Json<u8>,
-    json_vec: Json<Vec<u8>>,
-    json_ref: &Json<char>,
+    // json: Json<u8>,
+    // json_vec: Json<Vec<u8>>,
+    // json_ref: &Json<char>,
     fallible: Form<u64>,
 ) -> pavex_runtime::response::Response {
     todo!()
@@ -30,7 +30,7 @@ pub fn handler(
 
 pub fn blueprint() -> Blueprint {
     let mut bp = Blueprint::new();
-    bp.constructor(f!(crate::json), Lifecycle::RequestScoped);
+    // bp.constructor(f!(crate::json), Lifecycle::RequestScoped);
     bp.constructor(f!(crate::fallible), Lifecycle::RequestScoped)
         .error_handler(f!(crate::error_handler));
     bp.route(GET, "/home", f!(crate::handler));
